Migrate audio cue preferences to AndroidX Preference library
Move audio cue-related code from AudioCueSettingsActivity to a new AudioCueSettingsFragment class. When switching audio schemes, a fragment transaction is now used instead of using an Intent to (re)start a new AudioCueSettingsActivity instance.
